feat: New syscall: _kern_port_buffer_size_etc
This is a part of an attempt to get pkgman working. Right now,
pkgman checks for a "launch_daemon" port, which is managed by
"launch_daemon". Running /boot/system/servers/launch_daemon
for the first time shows that this syscall is missing.

After running the launch_daemon with this syscall implemented,
it crashes with a _kern_debugger, and with an error:
"looper must be locked before proceeding".