chore: stub _kern_open_attr_dir
Peacefully stub this syscall to get autoreconf working.

B_UNSUPPORTED is also quite common on Haiku, as Haiku doesn't implement
extended attributes support on many filesystems.

It is technically possible to implement this syscall through monika
or with the help of the loader, but it is quite complicated as we
have to wrap Linux's listxattr with a DIR*-style cursor API.