feat: uid/gid mapping
Added UID and GID mapping infrastructure in HyClone.
The user who started hyclone_server will be seen as the root (UID 0)
of HyClone.
Currently, adding additional maps is not supported yet. Users and
groups that are not mapped will simply have the same UID and GID
on HyClone as on the host.

This feature allows resource information such as team and file
metadata to be consistent with what getuid and getgid outputs.