feat: New syscall: _kern_install_team_debugger
This commit implements the basic debugger infrastructure on
HyClone.

Currently, only the "installation" process is implemented. This
means that the nub thread and some other objects essential for
debugging is created, but this nub thread cannot process any
message. Similarly, the debugger port cannot receive any message
(except for the ones generated by hyclone_server when a thread
and/or process dies).