fix: I found this tool on a random forum...
and my fate began to change.

Anime references aside, this commit introduces the use of ASAN in
`haiku_loader` and `hyclone_server`. (It is impossible to
introduce this is `monika` as it has to be as lightweight as
possible to be linked into `runtime_loader` and `libroot.so`, and
`monika` doesn't use `malloc` and similar C library functions
anyway).

Many memory-related bugs in these two components have been fixed.
This fixes the `pkgman` regression introduced by 8f99075 as well
as a long-standing bug where the screen would display
"id: cannot find name for user ID" on startup.