fix(packagefs): packagefs mount point
HyClone has long been under the false impression that `/boot`
is mounted as `packagefs`. The reality is only `/boot/system`
is `packagefs`, while `/boot` is some other filesystem.

This commit reflects the correct configuration. On HyClone,
`/boot` is mounted as `systemfs` pointing to its corresponding
path in the `hprefix`, while `/boot/system` is mounted as
`packagefs`.