fix(haiku_loader): Preserve emulated cwd
With `hyclone_server` supporting complicated hostfs-like mounts,
`vchroot_expand` may lose precision, especially when multiple
paths can be expanded to the same host path.

This also applies to the cwd. We therefore try to preserve it
by introducing the `--cwd` argument instead of relying on the
host to keep track of this.