fix(packagefs): Root directory id
The "root directory" of a Haiku `packagefs` volume is different
for each mount types.

For the default `/boot/system` (of type system), this directory is
the parent directory `/boot`, which is coincidentially also the
root of the boot volume. HyClone before this fix incorrectly
reported the root of the parent volume as the root directory of
every `packagefs` mount.

On Haiku, only one system and one user mount is allowed per
"root directory". Reporting the wrong root directory makes
mounting additional "type system" or "type home" volumes crash the
package daemon.