feat(hyclone_server): Directory watch
HyClone now supports watching directories.

Events attached to an emulated `IoContext` will be forwarded to
`VfsService`, which is then forwarded to `inotify` on Linux.

Currently, only directory-related events are processed. However,
this is enough for `package_daemon` to receive `B_ENTRY_CREATED`
and `B_ENTRY_REMOVED` for the `packages` folder and
install/uninstall packages on the fly.

HaikuPorter, 今は嘘が完全ですか？